+++ staging/src/k8s.io/apiserver/pkg/admission/plugin/webhook/errors/errors.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package config
+package errors
 
 import "fmt"
